Fetch constraint entities inline
