+++ engines/bops_api/app/jobs/bops_api/planning_application_dependency_job.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 
+require "faraday"
+require "uri"
+
 module BopsApi
   class PlanningApplicationDependencyJob < ApplicationJob
     queue_as :submissions
@@ -64,7 +67,7 @@ module BopsApi
           entities = designation.fetch(:entities, [])
 
           if entities.present?
-            BopsApi::FetchConstraintEntitiesJob.perform_later(planning_application_constraint, entities)
+            fetch_constraint_entities(planning_application_constraint, entities)
           end
         end
       end
@@ -141,6 +144,57 @@ module BopsApi
       end
     rescue ActiveRecord::RecordInvalid, NoMethodError => e
       Appsignal.send_error(e)
+    end
+
+    def fetch_constraint_entities(planning_application_constraint, entities)
+      planning_application_constraint.update!(data: fetch_entities(entities))
+    end
+
+    def fetch_entities(entities)
+      entities.map { |entity| fetch_entity(entity.fetch(:source)) }.compact
+    end
+
+    def fetch_entity(source)
+      return unless entity_url?(source)
+
+      uri = URI.parse(entity_url(source))
+
+      connection = Faraday.new(uri.origin) do |faraday|
+        faraday.response :raise_error
+        faraday.response :json, content_type: /\bjson$/
+      end
+
+      response = connection.get(uri.request_uri)
+
+      unless response.body.is_a?(Hash)
+        raise BopsApi::Errors::InvalidEntityResponseError, "Request for entity #{uri} returned a non-JSON response"
+      end
+
+      response.body
+    rescue Faraday::ResourceNotFound
+      nil
+    end
+
+    def entity_url(source)
+      case source
+      when Hash
+        "#{source.fetch(:url)}.json"
+      when String
+        "#{source}.json"
+      else
+        raise ArgumentError, "Invalid entity source: #{source.inspect}"
+      end
+    end
+
+    def entity_url?(source)
+      case source
+      when Hash
+        source.fetch(:url, nil).present?
+      when String
+        source.start_with?("https://www.planning.data.gov.uk/entity")
+      else
+        false
+      end
     end
   end
 end
